Fix: Resolve C language compilation warnings and improve code quality
This commit resolves various compilation warnings in the MEC system:

- Fixed memory management issues in memory_new.c
- Resolved string truncation warnings in monitor.c by using proper size limits
- Improved type safety in metrics.c and logging.c
- Enhanced thread safety in thread.c
- Fixed potential buffer overflow in config.c
- Updated monitor header to match Unix socket path limits

All fixes maintain functionality while eliminating compiler warnings.

+++ src/common/metrics.c
@@ -30,7 +30,7 @@ void metrics_report() {
         double avg_lat = (g_stats.frame_count > 0) ? (g_stats.total_latency_ms / g_stats.frame_count) : 0;
         
         LOG_INFO("PERF: FPS: %.2f | Avg Latency: %.3f ms | Frames: %ld", 
-                 fps, avg_lat, g_stats.frame_count);
+                 fps, avg_lat, (long)g_stats.frame_count);
     }
     pthread_mutex_unlock(&g_stats.lock);
 }
